ImageGrid: Refactor use of makeStyles to use of sx prop
