+++ src/components/ImageGrid/ImageGrid.tsx
@@ -1,4 +1,3 @@
-import { useStyles } from "./ImageGrid.css";
 import { useDispatch, useSelector } from "react-redux";
 import { Image } from "../../types/Image";
 import {
@@ -13,10 +12,12 @@ import {
   ImageList,
   ImageListItem,
   ImageListItemBar,
+  Box,
 } from "@mui/material";
 import { DropTargetMonitor, useDrop } from "react-dnd";
 import { NativeTypes } from "react-dnd-html5-backend";
-import { ImageIconLabel } from "./imageIconLabel";
+import { ImageIconLabel } from "./ImageIconLabel";
+import { Theme, styled } from "@mui/material/styles";
 
 type ImageGridProps = {
   onDrop: (item: { files: any[] }) => void;
@@ -55,8 +56,6 @@ export const ImageGrid = ({ onDrop }: ImageGridProps) => {
     }
   };
 
-  const classes = useStyles();
-
   const scaleFactor = useSelector(tileSizeSelector);
 
   const getSize = (scaleFactor: number) => {
@@ -72,22 +71,39 @@ export const ImageGrid = ({ onDrop }: ImageGridProps) => {
 
   const getSelectionStatus = (imageId: string) => {
     return selectedImages.includes(imageId)
-      ? classes.imageSelected
-      : classes.imageUnselected;
+      ? { border: "solid 2px blue", borderRadius: "3px" }
+      : { border: "none" };
   };
 
+  const StyledMain = styled("main")(({ theme }: { theme: Theme }) => ({
+    flexGrow: 1,
+    height: "100%",
+    paddingTop: theme.spacing(3),
+    marginLeft: theme.spacing(32),
+    transition: theme.transitions.create("margin", {
+      duration: theme.transitions.duration.enteringScreen,
+      easing: theme.transitions.easing.easeOut,
+    }),
+  }));
+
   return (
-    <main
+    <StyledMain
       ref={drop}
-      className={classes.main}
       style={{
         border: isOver ? "5px solid blue" : "",
       }}
     >
       <div>
-        <Container className={classes.container} maxWidth={false}>
+        <Container
+          sx={(theme) => ({
+            paddingBottom: theme.spacing(8),
+            paddingTop: theme.spacing(8),
+            height: "100%",
+          })}
+          maxWidth={false}
+        >
           <ImageList
-            className={classes.gridList}
+            sx={{ transform: "translateZ(0)", height: "100%" }}
             cols={Math.floor(6 / scaleFactor)}
             rowHeight={"auto"}
           >
@@ -96,9 +112,21 @@ export const ImageGrid = ({ onDrop }: ImageGridProps) => {
                 key={image.id}
                 onClick={() => onSelectImage(image)}
                 style={getSize(scaleFactor)}
-                className={getSelectionStatus(image.id)}
+                sx={getSelectionStatus(image.id)}
               >
-                <img alt="" src={image.src} className={classes.imageTile} />
+                <Box
+                  component="img"
+                  alt=""
+                  src={image.src}
+                  sx={{
+                    width: "100%",
+                    height: "100%",
+                    objectFit: "contain",
+                    top: 0,
+                    transform: "none",
+                  }}
+                />
+
                 <ImageListItemBar
                   position="top"
                   sx={{
@@ -114,6 +142,6 @@ export const ImageGrid = ({ onDrop }: ImageGridProps) => {
           <ImageGridAppBar />
         </Container>
       </div>
-    </main>
+    </StyledMain>
   );
 };
