+++ src/components/ImageGrid/ImageIconLabel.tsx
